Move packaged-path helper to shared for worker import boundary
